Feature/#127 dto update

+++ src/main/java/com/ssafy/ssashinsa/heyfy/exchange/dto/external/shinhan/ShinhanUpdateAccountRequestDto.java
@@ -12,6 +12,6 @@ public class ShinhanUpdateAccountRequestDto {
     private ShinhanCommonRequestHeaderDto Header;
     private String accountNo;
 
-    private Long transactionBalance;
-    private double transactionSummary;
+    private Double transactionBalance;
+    private Double transactionSummary;
 }
